feat: add ssh connection
Handle ssh connection from selected profile.
